requirements.txt
django
sentence-transformers
PyMuPDF
pytesseract
Pillow
openai